eulerian: first pass
Count the number of Eulerian paths on the 2d-dimensional hypercube.
This counts the number of ways the edges of the hypercube can be
directed such that every vertex has equal numbers of "in" and "out"
edges.

Naive first pass finds f(4) = 2970 in <0.01s, requiring 15979
recursions; f(6) estimated to require 2^37 minutes (about 260,000 years)
based on running for 1 minute.